Changed: Change how dropdowns with `ButtonView` and `SplitButtonView` are created.
